smart-crop-api/requirements.txt
Flask==3.1.0
Pillow==11.0.0
requests==2.32.3
gunicorn==23.0.0

